seperate dispay file
